refactor: removed eventsource initializer for rollback
