+++ sample-operators/tomcat-operator/src/test/java/io/javaoperatorsdk/operator/sample/TomcatOperatorE2E.java
@@ -11,11 +11,9 @@ import io.fabric8.kubernetes.client.ConfigBuilder;
 import io.fabric8.kubernetes.client.extended.run.RunConfigBuilder;
 import io.javaoperatorsdk.operator.Operator;
 import io.javaoperatorsdk.operator.config.runtime.DefaultConfigurationService;
-import io.javaoperatorsdk.operator.sample.tomcat.TomcatEventSourceInitializer;
 import io.javaoperatorsdk.operator.sample.tomcat.TomcatReconciler;
 import io.javaoperatorsdk.operator.sample.tomcat.resource.Tomcat;
 import io.javaoperatorsdk.operator.sample.tomcat.resource.TomcatSpec;
-import io.javaoperatorsdk.operator.sample.webapp.WebappEventSourceInitializer;
 import io.javaoperatorsdk.operator.sample.webapp.WebappReconciler;
 import io.javaoperatorsdk.operator.sample.webapp.resource.Webapp;
 import io.javaoperatorsdk.operator.sample.webapp.resource.WebappSpec;
@@ -42,13 +40,8 @@ public class TomcatOperatorE2E {
     if ("true".equals(System.getenv("RUN_OPERATOR_IN_TEST"))) {
       Operator operator = new Operator(client, DefaultConfigurationService.instance());
 
-      TomcatEventSourceInitializer tomcatEventSourceInitializer =
-          TomcatEventSourceInitializer.createInstance(client);
-      operator.register(new TomcatReconciler(client, tomcatEventSourceInitializer));
-
-      WebappEventSourceInitializer webappEventSourceInitializer =
-          WebappEventSourceInitializer.createInstance(client);
-      operator.register(new WebappReconciler(client, webappEventSourceInitializer));
+      operator.register(new TomcatReconciler(client));
+      operator.register(new WebappReconciler(client));
 
       operator.start();
     }
